Update LandingPage components with new logo, add NavBar, and enhance styling for improved responsiveness
